backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pymssql
openai==1.50.0
pydantic==2.9.0
httpx==0.27.0
